Tests for app links implemented

+++ controllers/appLinksController.js
@@ -108,7 +108,8 @@ class AppLinksController {
         return res.status(404).json({
           success: false,
           status: 404,
-          message: 'App not found'
+          message: 'App not found',
+          error: 'App not found'
         });
       }
 
@@ -140,29 +141,83 @@ class AppLinksController {
         return res.status(404).json({
           success: false,
           status: 404,
-          message: 'AppLinks not found'
+          message: 'Apps not found',
+          error: 'Apps not found'
         });
       }
 
-      const updatedAppLinks = await appLinks.updateAppLinks(id, name, url, icon, userId);
+      // Validate user input
+      if (!name || !url || !icon) {
+        return res.status(404).json({
+          success: false,
+          status: 404,
+          message: 'App Links name, icon and url required',
+          error: 'App Links name, icon and url required'
+        })
+      }
+
+      // check if app link id is valid
+      if (isNaN(id) || parseInt(id) <= 0) {
+        return res.status(400).json({
+          success: false,
+          status: 400,
+          message: "App id is not valid",
+          error: "App id is not valid"
+        });
+      }
+
+      // Check for invalid inputs
+      if (typeof name !== 'string' || typeof url !== 'string' || typeof icon !== 'string') {
+        return res.status(404).json({
+          success: false,
+          status: 404,
+          message: "App name, icon and url required, invalid inputs are not allowed",
+          error: "App name, icon and url required, invalid inputs are not allowed"
+        })
+      }
+
+      // check if app link already exists
+      const appExists = await appLinks.getAppLinksByName(name);
+      if (appExists && appExists.id !== parseInt(id)) {
+        return res.status(409).json({
+          success: false,
+          status: 409,
+          message: 'App already exists',
+          error: 'App already exists'
+        });
+      }
+
+      // check if user is the owner of the app
+      if (apps.user_id !== userId) {
+        return res.status(403).json({
+          success: false,
+          status: 403,
+          message: 'You are not allowed to update this app',
+          error: 'You are not allowed to update this app'
+        })
+      }
+
+      const updatedAppLinks = await appLinks.updateAppLinks(id, name, icon, url, userId);
+      
       return res.status(200).json({
         success: true,
         status: 200,
         message: 'App updated successfully',
-        data: [{
-          id: updatedAppLinks,
+        data: {
+          id: id,
           name,
           icon,
           url,
           user_id: userId
-        }]
+        }
       });
     } catch (error) {
-      console.error(error);
-      return res.status(500).json({
+      console.log(error);
+      return res.status(404).json({
         success: false,
-        status: 500,
-        message: 'Internal server error'
+        status: 404,
+        message: 'Apps not found',
+        error: 'Apps not found'
       });
     }
   }
@@ -181,11 +236,36 @@ class AppLinksController {
         });
       }
 
+      // check if app link id is valid
+      if (isNaN(id) || parseInt(id) <= 0) {
+        return res.status(400).json({
+          success: false,
+          status: 400,
+          message: "App id is not valid"
+        });
+      }
+
+      // check if user is the owner of the app
+      if (apps.user_id !== req.user.id) {
+        return res.status(403).json({
+          success: false,
+          status: 403,
+          message: 'You are not allowed to delete this app'
+        })
+      }
+
       await appLinks.deleteAppLinks(id);
       return res.status(200).json({
         success: true,
         status: 200,
-        message: 'AppLinks deleted successfully'
+        message: 'App deleted successfully',
+        data: {
+          id: id,
+          name: apps.name,
+          icon: apps.icon,
+          url: apps.url,
+          user_id: apps.user_id
+        }
       });
     } catch (error) {
       console.error(error);
